src/test/java/ch/supsi/dti/isin/benchmark/adapter/consistenthash/memento/MementoFactoryTests.java
package ch.supsi.dti.isin.benchmark.adapter.consistenthash.memento;


import static org.junit.jupiter.api.Assertions.assertNotNull;

import java.util.function.Supplier;

import org.junit.jupiter.api.DisplayNameGeneration;
import org.junit.jupiter.api.DisplayNameGenerator;
import org.junit.jupiter.api.Test;

import ch.supsi.dti.isin.benchmark.adapter.ConsistentHashFactoryContract;
import ch.supsi.dti.isin.benchmark.config.AlgorithmConfig;
import ch.supsi.dti.isin.consistenthash.memento.MementoEngine;

/**
 * Suite to test the {@link MementoFactory} class.
 * 
 * @author Massimo Coluzzi
 */
@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
public class MementoFactoryTests implements ConsistentHashFactoryContract<MementoFactory>
{


    /* ******************* */
    /*  INTERFACE METHODS  */
    /* ******************* */


    /**
     * {@inheritDoc}
     */
    @Override
    public MementoFactory sampleValue( AlgorithmConfig config )
    {

        return new MementoFactory( config );

    }


    /* ************** */
    /*  TEST METHODS  */
    /* ************** */


    @Test
    public void default_configuration_should_return_default_engine()
    {

        final MementoFactory factory = sampleValue( CONFIG );
        
        final Supplier<MementoEngine> supplier = factory.createEngineInitializer( FUNCTION, NODES );
        assertNotNull( supplier );

        final MementoEngine engine = supplier.get();
        assertNotNull( engine );

    }

}
